feat(cli): added a version-command command to CLI
- Moved index.js to cli.js
- Created a new index.js that exposes library functions
- Moved configuration initialization logic to its own file
- Added a new 'next-version' command that returns the result of the
  pre() function. This could be used in other scripts. e.g.:

  `npm version $(semantic-release next-version --type)`
  `semantic-release next-version)`
  `semantic-release next-version --type)`
  `semantic-release next-version --version)`
